Fixed hint bug that caused there to be no available hints in certain scenarios.
Example: there are three findable object. a hint is given for one object, but the other two objects are found and not the hinted object. when another hint is requested, the record of hints given still has the one lingering object within it, so the list of available hints is empty. before asking for another hint, we need to update the hints given list, if available. in this case, it would mean reseting it to be empty.